--- src/main/java/com/example/sopkathon/domian/ping/dto/PingListRes.java
@@ -1,4 +1,0 @@
-package com.example.sopkathon.domian.ping.dto;
-
-public record PingListRes() {
-}
