+++ src/main/java/com/example/sopkathon/domian/ping/dto/req/PingRequest.java
@@ -1,4 +1,4 @@
-package com.example.sopkathon.domian.ping.dto;
+package com.example.sopkathon.domian.ping.dto.req;
 
 public record PingRequest(String situation, String ping) {
 }
